pyproject.toml
[tool.poetry]
name = "routerus"
version = "0.5.1"
description = "VPN Management System with Shadowsocks and Xray Integration"
authors = ["Anfikus <anfisa.kov@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
bcrypt = "^4.2.0"
cffi = "^1.17.1"
click = "^8.1.7"
cryptography = "^43.0.3"
cython = "^3.0.11"
django = "^5.1.3"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
numpy = "^2.1.3"
packaging = "^24.2"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
pycparser = "^2.22"
pynacl = "^1.5.0"
pyopenssl = "^24.2.1"
pypng = "^0.20220715.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.2"
pyyaml = "^6.0.2"
qrcode = "^8.0"
shadowsocks = "^2.8.2"
six = "^1.16.0"
sqlparse = "^0.5.1"
subnet-utils = "^0.5.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
wireguard = "^1.0.0"
xray = "^0.7.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
black = "^23.3b0"
flake8 = "^6.0.0"
isort = "^5.10.1"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
